+++ app/src/main/java/com/juniori/puzzle/ui/home/HomeViewModel.kt
@@ -13,6 +13,7 @@ import com.juniori.puzzle.data.weather.WeatherRepository
 import com.juniori.puzzle.domain.usecase.GetUserInfoUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
+import java.util.*
 import javax.inject.Inject
 
 @HiltViewModel
@@ -36,9 +37,8 @@ class HomeViewModel @Inject constructor(
     private val _weatherList = MutableLiveData<List<WeatherItem>>(emptyList())
     val weatherList: LiveData<List<WeatherItem>> = _weatherList
 
-    private val _weatherMainList = MutableLiveData<WeatherItem>().apply {
-        WeatherItem("", "", 0, 0, 0, 0, "", "")
-    }
+    private val _weatherMainList =
+        MutableLiveData(WeatherItem(Date(), 0, 0, 0, 0, "", ""))
     val weatherMainList: LiveData<WeatherItem> = _weatherMainList
 
     fun setDisplayName() {
