+++ app/src/main/res/layout/item_weather_detail.xml
@@ -20,7 +20,7 @@
             style="@style/Theme.Puzzle.TextSmall.Bold"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@{item.time}"
+            setTime="@{item.date}"
             app:layout_constraintBottom_toTopOf="@id/weather_icon_iv"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
